+++ vasilievGavril/package.json
@@ -9,6 +9,7 @@
     "web": "expo start --web"
   },
   "dependencies": {
+    "@react-navigation/bottom-tabs": "^6.5.4",
     "@react-navigation/native": "^6.1.3",
     "@react-navigation/native-stack": "^6.9.9",
     "expo": "~47.0.12",
@@ -16,7 +17,7 @@
     "react": "18.1.0",
     "react-native": "0.70.5",
     "react-native-safe-area-context": "^4.5.0",
-    "react-native-screens": "^3.19.0"
+    "react-native-screens": "^3.20.0"
   },
   "devDependencies": {
     "@babel/core": "^7.12.9"
